+++ RELEASE_NOTES.md
@@ -1,2 +1,1 @@
-* Fixed YACL compatibility, as it was not usable with a controller before.
-* Registered all supported mod mixins.
+* Fix missing texture on custom radial menu icons